Refactor recent requests table logic to new module
Moved recent requests table and pagination logic from report.mjs to a new report.trackingRecords.mjs module. Updated report.mjs to import and use loadRecentRequestsTable from the new module for better separation of concerns and maintainability.
